netdev-dpdk: add dpdk vhost ports
This patch adds support for a new port type to userspace datapath
called dpdkvhost. This allows KVM (QEMU) to offload the servicing
of virtio-net devices to its associated dpdkvhost port. Instructions
for use are in INSTALL.DPDK.

This has been tested on Intel multi-core platforms and with clients
that have virtio-net interfaces.

 ver 6:
   - rebased with master
   - modified to use DPDK v1.8.0 vhost library
   - reworked for review comments
 ver 5:
   - rebased against latest master
 ver 4:
   - added eventfd_link.h and eventfd_link.c to EXTRA_DIST in
 utilities/automake.mk
   - rebased with master to work with DPDK 1.7 ver 3:
   - rebased with master
 ver 2:
   - rebased with master

Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Maryam Tahhan <maryam.tahhan@intel.com>