ovn-nbd: Match bindings on lport name.
When ovn-nbd is looking through bindings to figure out if any logical
ports in the northbound DB need to have their 'up' state changed, it
should be using the logical port's name instead of UUID for matching
up to rows in the OVN db.

The switch to using the name instead of UUID means we don't have a
quick lookup so we use the same approach that's used for handling
updates in the other direction.  We start by hashing all of the
logical ports by name.  As we iterate through the bindings, we can
then look up ports by name quickly.

Signed-off-by: Russell Bryant <rbryant@redhat.com>