ovn: Rename ovn-nbd to ovn-northbound.
It came up in a recent discussion that the use of 'nbd' could cause
confusion since nbd is used for "Network Block Device", which may be
used in many of the same environments that OVN would be used in.

To avoid the possible issue, just rename ovn-nbd to ovn-northbound.
While we're at it, create a subdirectory under ovn/ to hold all files
related to ovn-northbound.  For now it's just a single C file, but we
may add more source files in the future.

Signed-off-by: Russell Bryant <rbryant@redhat.com>