ovn-controller: Fix invalid br_int_name handling.
While doing some testing, I noticed the following error message:

  Integration bridge '<garbage>' dissapeared

The reason is that the code kept around the value of the "ovn-bridge"
configuration without copying it.  The result was pointing to bogus
memory.  You would only see this if you set "ovn_bridge".  If you
relied on the default, the bug would not occur.

Signed-off-by: Russell Bryant <rbryant@redhat.com>