ovn: Add "localnet" logical port type.
Introduce a new logical port type called "localnet".  A logical port
with this type also has an option called "network_name".  A "localnet"
logical port represents a connection to a locally accessible network.
ovn-controller will use the ovn-bridge-mappings configuration to
figure out which patch port on br-int should be used for this port.

Signed-off-by: Russell Bryant <rbryant@redhat.com>