python: Set up initial tox test environment.
tox is a tool for running Python tests in virtual Python environments.
This initial setup only attempts to install the ovs Python library in
both a virtual Python 2.7 environment and a virtual Python 3.4
environment.  We also run the beginning of a test suite, which only
imports the code.  This quickly reproduces the problem that the library
fails to install for Python 3.  Future patches will start addressing
Python 3 compatibility.

Signed-off-by: Russell Bryant <russell@ovn.org>