python: Resolve pep8 comparison errors.
Resolve pep8 errors:

  E711 comparison to None should be 'if cond is None:'

The reason comparing against None with "is None" is preferred over
"== None" is because a class can define its own equality operator and
produce bizarre unexpected behavior.  Using "is None" has a very
explicit meaning that can not be overridden.

  E721 do not compare types, use 'isinstance()'

This one is actually a mistake by the tool.  The line of code looks like
it's using the types module from the Python stdlib, but it's actually
ovs.db.types.  Change the import to make this more clear and eliminate
the error.

Signed-off-by: Russell Bryant <russell@ovn.org>