xml2nroff: Make compatible with Python 3.
Update xml2nroff to work with either Python 2 or Python 3.  The changes
are minor and can be managed without the use of the six library.

For iterating a dict, we now use:

    for k, v in subst.items():

This is the same as .iteritems() in Python 2.  It's technically a little
less efficient for Python 2, as items() returns a list instead of an
iterator in Python 2.  We're dealing with a very small set of items here
so it really doesn't matter.

Signed-off-by: Russell Bryant <russell@ovn.org>