python: Add basic testing of ovs.json.
Add basic testing of the normal code path in the JSON parser.

This at least shows how to add more Python unit tests that validate the
code works on multiple Python versions.  For things that are relatively
simple and Python specific, this seems like an easier way to add test
cases than autotest.

Signed-off-by: Russell Bryant <russell@ovn.org>