python: Run ovsdb-doc from tox.
The ovsdb-doc utility uses the ovs Python library.  This patch runs it
in the virtual Python environments (2.7 and 3.4) to ensure that it
completes successfully.  This helped quickly expose some more bugs in
the Python 3 port.

The patch includes the addition of the -s/--silent option to ovsdb-doc,
as we don't want the nroff output sent to stdout in this case.

Signed-off-by: Russell Bryant <russell@ovn.org>