ovn: Add hostname to Chassis.
We currently use the system-id from the Open_vSwitch schema on each host
to populate the unique name field of a Chassis in OVN_Southbound.  On
most systems, this is a UUID.  It would be very convenient to also have
the hostname available as that will allow people to more quickly
identify which host a Chassis record is associated with in most cases.

This is also useful for correlating an OVN Chassis with the knowledge of
hosts that exists in other systems (such as OpenStack).

Also add the hostname of each chassis to the output of "ovn-sbctl show".

Signed-off-by: Russell Bryant <russell@ovn.org>