ovn: Add --foreground to ovn-ctl.
Add an option to ovn-ctl to keep the daemon it's starting in the
foreground.

OpenStack has a development environment called "devstack".  We've been
updating our devstack integration to use ovn-ctl instead of starting
all daemons manually, but devstack likes to run daemons in the
foreground in a bunch of tabs in a screen session.

Signed-off-by: Russell Bryant <russell@ovn.org>