WIP: Limited general VXLAN support for OVN.
This is an attempt at hacking in general VXLAN support for OVN.
With VXLAN, we only send the network ID and not the source and destintaion
logical port IDs.

source logical port - we just accept that we don't have it in the egress
pipeline.  The only limitation this imposes is that we can't match on
"inport" in egress ACLs.

destination logical port - we re-do the l2_lkup stage that happens at the
end of the ingress pipeline at the beginning of the egress pipeline.  It
will be skipped if we received the destination as metadata.

TODO:
 - This is completely untested
 - lacking documentation updates in several places, in and out of code