ovn: Discard flows for non-local ports.
Discard some OpenFlow flows that will never match.  This includes
flows that match on a non-local inport in the ingress pipeline or a
non-local outport in the egress pipeline of a logical switch.

This is most useful for networks with a large number of ports or ACLs
that use large address sets.

Signed-off-by: Russell Bryant <russell@ovn.org>
Tested-by: Miguel Angel Ajo Pelayo <majopela@redhat.com>